Move base classes to separate files
